Add foldreads2.
This version requires adapter sequences for the Y-adapter, which are used
to explicitly exclude putative Y-Y molecules.  Read pairs with hairpin
sequence in a different position in read 1 and read 2 are also excluded.

Short Y-hairpin molecules may contain useful information 3' of the hairpin
that was previously excluded. If available, we now use this information.
We recover only an additional 0.7% of reads, however single base errors and
quality scores are improved throughout.